# backend/docs/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Page_AllocationListItem_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AllocationListItem"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "pageSize": {
            "title": "Pagesize",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          },
          "totalPages": {
            "title": "Totalpages",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "page",
          "pageSize",
          "total",
          "totalPages"
        ],
        "title": "Page[AllocationListItem]",
        "type": "object"
      },
...
      "WithdrawInput": {
        "description": "An Administrator withdrawing an offer before the trainer answers.",
        "properties": {
          "reason": {
            "maxLength": 2000,
            "minLength": 10,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "reason"
        ],
        "title": "WithdrawInput",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "description": "JWT access token.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "email": "ict@upf.go.ug",
      "name": "Uganda Police Force — Directorate of Human Resource Development"
    },
...
    "license": {
      "name": "Restricted — Uganda Police Force internal use"
    },
    "title": "Trainer Prediction System API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/allocations": {
      "get": {
        "description": "Every approved assignment, filterable by status, programme, trainer, approving officer, and approval date range.",
        "operationId": "list_allocations_api_v1_allocations_get",
        "parameters": [
          {
            "description": "Allocation status.",
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Allocation status.",
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "programmeId",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Programmeid"
            }
          },
          {
            "in": "query",
            "name": "trainerId",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Trainerid"
            }
          },
          {
            "in": "query",
            "name": "approvedBy",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Approvedby"
            }
          },
          {
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "pageSize",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Pagesize",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "sortBy",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sortby"
            }
          },
          {
            "in": "query",
            "name": "sortDir",
            "required": false,
            "schema": {
              "default": "asc",
              "pattern": "^(asc|desc)$",
              "title": "Sortdir",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_AllocationListItem_"
                }
              }
            },
            "description": "A page of allocations."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List allocations",
        "tags": [
          "Allocations"
        ]
      },
      "post": {
        "description": "**This is the decision.** BR-02 restricts it to a Training Administrator and BR-06 requires it to be explicit — there is no auto-approve and no bulk approve anywhere in this API.\n\nIn one transaction the server verifies the candidate comes from the current ranking, **re-checks the hard gates against live data** (a trainer may have become unavailable since the run — 409 if so), freezes the score, breakdown, rank, weights and rationale, draws a registry number, sets the allocation `PENDING_TRAINER`, moves the programme to `AWAITING_RESPONSE`, notifies the trainer, and writes the audit entry (BR-07).\n\n`weights` and `weightsWereSimulated` in the body are accepted for compatibility and **ignored** — the frozen weights come from the run itself, because a receipt whose figures were supplied by the client could be made to say anything.",
        "operationId": "approve_api_v1_allocations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveAllocationInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AllocationListItem"
                }
              }
            },
            "description": "Approved. The Decision Receipt is the response body."
          },
          "403": {
            "description": "Only a Training Administrator may approve (BR-02)."
          },
          "404": {
            "description": "No such candidate."
          },
          "409": {
            "description": "The ranking is superseded, the requirements changed since it was generated, the programme already has an allocation, or the candidate no longer passes the gates."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Approve an allocation (FR-08)",
        "tags": [
          "Allocations"
        ]
      }
    },
    "/api/v1/allocations/{allocation_id}": {
      "get": {
        "description": "The frozen snapshot: score, breakdown, rank and weights **as they stood at approval**, the rationale the trainer was shown, the approving officer with their rank, the registry number, the trainer's response, and the linked evaluation if one exists.\n\nNone of it is recomputed. An evaluation recorded next month changes future rankings; it must not change the justification for a decision taken today.",
        "operationId": "get_allocation_api_v1_allocations__allocation_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "allocation_id",
            "required": true,
            "schema": {
              "title": "Allocation Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AllocationListItem"
                }
              }
            },
            "description": "The receipt."
          },
          "403": {
            "description": "A Trainer may only view their own allocation."
          },
          "404": {
            "description": "No such allocation."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "The Decision Receipt",
        "tags": [
          "Allocations"
        ]
      }
    },
    "/api/v1/allocations/{allocation_id}/mark-conducted": {
      "post": {
        "description": "Sets the allocation to `CONDUCTED` and the programme with it. **This is the gate that unlocks FR-10** — an evaluation cannot be recorded against a course nobody has confirmed happened.",
        "operationId": "mark_conducted_api_v1_allocations__allocation_id__mark_conducted_post",
        "parameters": [
          {
            "in": "path",
            "name": "allocation_id",
            "required": true,
            "schema": {
              "title": "Allocation Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AllocationListItem"
                }
              }
            },
            "description": "Marked conducted."
          },
          "409": {
            "description": "The trainer has not confirmed this assignment."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Record that the training took place",
        "tags": [
          "Allocations"
        ]
      }
    },
    "/api/v1/allocations/{allocation_id}/promote-next": {
      "post": {
        "description": "Takes the next candidate **from the same existing ranking**. FR-08 requires that a decline does not trigger a new prediction, so `reusedExistingRun` is always `true` and the frontend states it to the user.\n\nCandidates passed over on the way — already allocated, or no longer through the gates — are listed in `skipped` and audited individually as `CANDIDATE_SKIPPED`.",
        "operationId": "promote_next_api_v1_allocations__allocation_id__promote_next_post",
        "parameters": [
          {
            "in": "path",
            "name": "allocation_id",
            "required": true,
            "schema": {
              "title": "Allocation Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PromoteNextResponse"
                }
              }
            },
            "description": "The next candidate now holds the offer."
          },
          "409": {
            "description": "The allocation was not declined, a replacement was already promoted, or no remaining candidate meets the requirements."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Promote the next-ranked candidate after a decline (FR-08)",
        "tags": [
          "Allocations"
        ]
      }
    },
    "/api/v1/allocations/{allocation_id}/withdraw": {
      "post": {
        "description": "Requires a reason, which is recorded and sent to the trainer. The programme returns to `PREDICTED`, so another candidate can be approved from the same ranking without a re-run.",
        "operationId": "withdraw_api_v1_allocations__allocation_id__withdraw_post",
        "parameters": [
          {
            "in": "path",
            "name": "allocation_id",
            "required": true,
            "schema": {
              "title": "Allocation Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WithdrawInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AllocationListItem"
                }
              }
            },
            "description": "Withdrawn."
          },
          "409": {
            "description": "The trainer has already responded."
          },
          "422": {
            "description": "No reason was given."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Withdraw an offer before the trainer answers",
        "tags": [
          "Allocations"
        ]
      }
    },
...
    "/api/v1/reference/all": {
      "get": {
        "description": "Returns all ten reference lists together. Preferred over the individual endpoints when a screen needs more than one — the programme form needs six, and one request is better than six.",
        "operationId": "all_reference_api_v1_reference_all_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReferenceBundle"
                }
              }
            },
            "description": "Every reference list."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Every lookup list in one response",
        "tags": [
          "Reference data"
        ]
      }
    },
...
    "/api/v1/trainers/me/assignments": {
      "get": {
        "description": "Pending invitations, confirmed upcoming courses, and past assignments.\n\nEach pending item carries `frozenRationale` — **the trainer is entitled to see why they were selected**, not merely that they were.",
        "operationId": "my_assignments_api_v1_trainers_me_assignments_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssignmentsResponse"
                }
              }
            },
            "description": "The caller's own assignments."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "My assignments",
        "tags": [
          "Trainer assignments"
        ]
      }
    },
    "/api/v1/trainers/me/assignments/{allocation_id}/accept": {
      "post": {
        "description": "Confirms the assignment, records the response time, moves the programme to `ALLOCATED`, sets the trainer `ASSIGNED` — which caps their AVAILABILITY score on future runs, so the system stops favouring someone already committed — and notifies the approving Administrator.",
        "operationId": "accept_api_v1_trainers_me_assignments__allocation_id__accept_post",
        "parameters": [
          {
            "in": "path",
            "name": "allocation_id",
            "required": true,
            "schema": {
              "title": "Allocation Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AllocationListItem"
                }
              }
            },
            "description": "Confirmed."
          },
          "403": {
            "description": "That assignment belongs to another trainer."
          },
          "409": {
            "description": "It is no longer awaiting a response."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Accept an assignment (FR-09)",
        "tags": [
          "Trainer assignments"
        ]
      }
    },
...
  },
...
}